Read and process mulitple data files

Apply spectral binning on multiple data files.

readFiles(files, dp, scans)

Arguments

files
A vector of converted data file paths

dp
An integer denoting the number of decimal places for spectral binning

scans
A vector of scan numbers that should be retrieved

Value

A list containing peak lists for the relevant scans with combined scan ranges for each present mode in the data file.

Details

Parallel processing is managed by the future package. This can be specified using the plan() function. See the example below and ?future::plan for details on how this can be specified.

Examples

## Example file path
file_paths <- metaboData::filePaths('FIE-HRMS',
                                    'BdistachyonEcotypes')[1]

## Optionally declare parallel processing backend
# plan(future::multisession,workers = 2)
                                                                        
## Process example file
res <- readFiles(file_paths,
                 dp = 2,
                 scans = 6:17)